Real esm module with pure typescript FTW
