+++ jest.config.json
@@ -21,5 +21,8 @@
         "useESM": true
       }
     ]
+  },
+  "moduleNameMapper": {
+    "^(\\.{1,2}/.*)\\.js$": "$1"
   }
 }
