+++ src/Grid/Grid.spec.js
@@ -2,7 +2,7 @@ import React from 'react'
 import renderer from 'react-test-renderer'
 import 'jest-styled-components'
 
-import Grid from './index.tsx'
+import Grid from './index.js'
 
 import styled from 'styled-components'
