+++ src/Search/Search.stories.tsx
@@ -1,8 +1,9 @@
 import React from 'react'
-import Search from './'
 import { Story } from '@storybook/react'
 import styled from 'styled-components'
 import { action } from '@storybook/addon-actions'
+
+import Search from './index.js'
 
 export default {
   title: 'Components/Search',
