+++ src/index.ts
@@ -1,11 +1,11 @@
-export { default as TestComponent } from './TestComponent'
-export { default as Grid } from './Grid'
-export { default as Drawer } from './Drawer'
-export { default as Search } from './Search'
-export { default as ListCard } from './ListCard'
-export { default as DropDown } from './Dropdown'
-export { default as AppBar } from './AppBar'
-export { default as Map } from './Map'
-export * from './Heading'
-export { default as Section } from './Section'
-export { default as Dialog } from './Dialog'
+export { default as TestComponent } from './TestComponent/index.js'
+export { default as Grid } from './Grid/index.js'
+export { default as Drawer } from './Drawer/index.js'
+export { default as Search } from './Search/index.js'
+export { default as ListCard } from './ListCard/index.js'
+export { default as DropDown } from './Dropdown/index.js'
+export { default as AppBar } from './AppBar/index.js'
+export { default as Map } from './Map/index.js'
+export * from './Heading/index.js'
+export { default as Section } from './Section/index.js'
+export { default as Dialog } from './Dialog/index.js'
